De-abstract main code, prevent duplicate status messages.
